// src/store/modules/user/types.ts
export type RoleType = '' | '*' | 'admin' | 'user'
export interface UserState {
  address: string
  avatar: string
  bgImage: string
  city: string
  country: string
  createTime: string
  distinct: string
  email: string
  nickname: string
  phone: string
  province: string
  sex: string
  state: string
  token: string
  userId: string
  username: string
}